refactor(client): 抽取宿主推送与自动激活触发器到 sidebar/use-host-feeds.ts
Sidebar.tsx 拆分第三块：agent-terminals / agent-opens 两条 WS 推送
（含重连退避与失败上限）、子代理自动激活（去抖重估）、后台任务自动
激活、拓扑跳回（subagentJumpRef）连同 FAILURE_LIMIT /
AUTO_OPEN_DEBOUNCE_MS 常量原样迁出，注释逐字保留。Sidebar 侧仅
消费返回的 subagentJumpRef（renderTab 的 onSubagentJump 装填）。

行为不变：五个 effect 的依赖数组与闭包值逐字相同，hook 每 render
无条件调用，effect 挂载顺序保持（feeds → center → float →
bottomAuto → drag）。
